to float16

+++ transform/eigen_posenc.py
@@ -32,7 +32,7 @@ def compute_posenc_stats(data, is_undirected):
       edge_weight = edge_weight.to(device='cuda')
       
       # Create dense Laplacian matrix
-      L = torch.zeros((N, N), device='cuda')
+      L = torch.zeros((N, N), device='cuda', dtype=torch.float16))
       L[edge_index[0], edge_index[1]] = edge_weight
       
       # Compute eigenvalues and eigenvectors
